Pull getURLToSubmitTo out for reuse
