+++ collect_app/src/main/java/org/odk/collect/android/upload/InstanceServerUploader.java
@@ -1,7 +1,10 @@
 package org.odk.collect.android.upload;
 
 import android.content.ContentValues;
+import android.content.SharedPreferences;
 import android.net.Uri;
+import android.preference.PreferenceManager;
+import android.support.annotation.NonNull;
 
 import com.google.android.gms.analytics.HitBuilders;
 
@@ -10,6 +13,7 @@ import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.dto.Instance;
 import org.odk.collect.android.http.HttpHeadResult;
 import org.odk.collect.android.http.OpenRosaHttpInterface;
+import org.odk.collect.android.preferences.PreferenceKeys;
 import org.odk.collect.android.provider.InstanceProviderAPI;
 import org.odk.collect.android.upload.result.SubmissionUploadAuthRequested;
 import org.odk.collect.android.upload.result.SubmissionUploadFatalError;
@@ -22,8 +26,10 @@ import org.odk.collect.android.utilities.WebCredentialsUtils;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
 import java.net.URI;
 import java.net.URLDecoder;
+import java.net.URLEncoder;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -33,6 +39,7 @@ import javax.net.ssl.HttpsURLConnection;
 import timber.log.Timber;
 
 public class InstanceServerUploader {
+    private static final String URL_PATH_SEP = "/";
     private static final String FAIL = "Error: ";
 
     private final OpenRosaHttpInterface httpInterface;
@@ -251,6 +258,61 @@ public class InstanceServerUploader {
         }
 
         return files;
+    }
+
+
+    /**
+     * Returns the URL this instance should be submitted to with appended deviceId.
+     *
+     * If the upload was triggered by an external app and specified an override URL, use that one.
+     * Otherwise, use the submission URL configured in the form
+     * (https://opendatakit.github.io/xforms-spec/#submission-attributes). Finally, default to the
+     * URL configured at the app level.
+     */
+    @NonNull
+    public String getURLToSubmitTo(Instance currentInstance, String deviceId, String overrideURL) {
+        String urlString;
+
+        if (overrideURL != null) {
+            urlString = overrideURL;
+        } else if (currentInstance.getSubmissionUri() != null) {
+            urlString = currentInstance.getSubmissionUri().trim();
+        } else {
+            urlString = getServerSubmissionURL();
+        }
+
+        // add deviceID to request
+        try {
+            urlString += "?deviceID=" + URLEncoder.encode(deviceId != null ? deviceId : "", "UTF-8");
+        } catch (UnsupportedEncodingException e) {
+            Timber.i(e, "Error encoding URL for device id : %s", deviceId);
+        }
+
+        return urlString;
+    }
+
+    private String getServerSubmissionURL() {
+
+        Collect app = Collect.getInstance();
+
+        SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(
+                Collect.getInstance());
+        String serverBase = settings.getString(PreferenceKeys.KEY_SERVER_URL,
+                app.getString(R.string.default_server_url));
+
+        if (serverBase.endsWith(URL_PATH_SEP)) {
+            serverBase = serverBase.substring(0, serverBase.length() - 1);
+        }
+
+        // NOTE: /submission must not be translated! It is the well-known path on the server.
+        String submissionPath = settings.getString(PreferenceKeys.KEY_SUBMISSION_URL,
+                app.getString(R.string.default_odk_submission));
+
+        if (!submissionPath.startsWith(URL_PATH_SEP)) {
+            submissionPath = URL_PATH_SEP + submissionPath;
+        }
+
+        return serverBase + submissionPath;
     }
 
     private void saveSuccessStatusToDatabase(Instance instance) {
